+++ common/tools/dev-tool/src/util/admin/migrate-package/codemods/replaceSinonStubs.ts
